IntegrationTests: add  AbstractKafkaClient
